he terminado la funcionalidad de agregar una animacion a todos los inputs del formulario de login
